+++ fiddles/angular2/fiddle-0044-GridStateRouting/app/components/entities/dictionary.spec.ts
@@ -1,0 +1,14 @@
+import {
+  describe,
+  ddescribe,
+  expect,
+  iit,
+  it
+} from '@angular/core/testing';
+import {Dictionary} from './dictionary';
+
+describe('Dictionary', () => {
+  it('should create an instance', () => {
+    expect(new Dictionary()).toBeTruthy();
+  });
+});