+++ fiddles/angular2/fiddle-0044-GridStateRouting/app/content/content.routes.ts
@@ -1,0 +1,27 @@
+import { RouterConfig }          from '@angular/router';
+import { ContentComponent } from './content.component';
+import { ContentDetailComponent } from './content-detail/content-detail.component';
+
+
+export const ContentRoutes:RouterConfig = [
+    {
+        path: '',
+        redirectTo: '/',
+        terminal: true
+    },
+    {
+        path: '',
+        redirectTo: '?id=report-a.csv',
+        terminal: true
+    },
+    {
+        path: '',
+        component: ContentComponent,
+        children: [
+            {
+                path: ':id',
+                component: ContentDetailComponent
+            }
+        ]
+    }
+];