+++ fiddles/angular2/fiddle-0044-GridStateRouting/app/meta.ts
@@ -1,0 +1,13 @@
+var metadata = {
+    fiddleHeader: 'Angular 2 - Grid State Routing',
+    urls: {
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0044-GridStateRouting'
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s',
+    onError: function(err) {
+        console.log(err);
+        document.write('<H3>' + err + '</H3>');
+    }
+}
+
+